Start extracting a base class for error handlers.
Use a template method pattern to make error handlers consistent, but
provide a way to customize the displayed output based on the
environment.
